bullets come out, but don't move. also, dropped the idea of infinite looping, which consumed too muchh CPU

+++ source/global.h
@@ -2,8 +2,8 @@
 	#define __GLOBAL_H__
 
 	typedef struct Point {
-		int32_t y;
-		int32_t x;
+		int16_t y;
+		int16_t x;
 	} Point;
 
 #endif //__GLOBAL_H__
